+++ examples/net/https_server.rb
@@ -3,12 +3,20 @@
 require 'modulation'
 require 'json'
 require 'localhost/authority'
+require 'time'
 
 Server = import('../../lib/nuclear/http/server')
 
+H_SERVER = 'Server'
+H_DATE = 'Date'
+
+SERVER_NAME = 'nuclear/0.6'
+HELLO_WORLD = 'Hello World'
+
 server = Server.new do |request, response|
-  response.write_head(200, 'Content-Type': 'application/json')
-  response.finish(request.inspect)
+  response.write_head(200, 
+    H_SERVER => SERVER_NAME, H_DATE => Time.now.httpdate)
+  response.finish(HELLO_WORLD)
 end
 
 authority = Localhost::Authority.fetch