+++ lib/nuclear.rb
@@ -9,10 +9,11 @@ Nuclear = import('./nuclear/core')
 module Nuclear
   VERSION = '0.6'
 
-  Promise     = import('./nuclear/core/promise')
-  IO          = import('./nuclear/io')
-  LineReader  = import('./nuclear/line_reader')
-  Net         = import('./nuclear/net')
-  Stream      = import('./nuclear/stream')
-  Thread      = import('./nuclear/thread')
+  Promise       = import('./nuclear/core/promise')
+  IO            = import('./nuclear/io')
+  LineReader    = import('./nuclear/line_reader')
+  Net           = import('./nuclear/net')
+  Stream        = import('./nuclear/stream')
+  Thread        = import('./nuclear/thread')
+  ResourcePool  = import('./nuclear/resource_pool')
 end